Move read more component into general folder
